requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# LLM, RAG, and Vector Store
langchain
langchain-google-genai
langchain-community
sentence-transformers
qdrant-client

# Database Connector (Penting untuk NL-to-SQL)
SQLAlchemy
PyMySQL

# Opsional: Untuk environment variables
python-dotenv

# read yml
PyYAML

# Data Manipulation
pandas
sqlparse

# Rate limiting & templating
slowapi
Jinja2

# Token counting & multi-LLM support
tiktoken
google-generativeai
transformers
httpx

# Opsional: Logging & debugging
loguru
rich